mod/guest: plucking helper functions into a separate `utils.rs`
